all commit

+++ database/migrations/2024_05_07_104136_create_registrations.php
@@ -14,7 +14,7 @@ return new class extends Migration
         Schema::create('registrations', function (Blueprint $table) {
             $table->id();
             $table->date('send_date');
-            $table->integer('cso_id');
+            $table->unsignedBigInteger('cso_id');
             $table->foreign('cso_id')->references('id')->on('csos')->cascadeOnDelete();
             $table->timestamps();
         });
